Refacto VideoInformationPane
