+++ web/src/collection/components/VideoDetailsPage/VideoInformationPane.js
@@ -12,53 +12,17 @@ import ExifPanel from "./ExifPanel";
 import ObjectsPanel from "./ObjectsPanel";
 import FileInfoPanel from "./FileInfoPanel";
 import { useIntl } from "react-intl";
+import VideoInformation from "./VideoInformation";
 
 const useStyles = makeStyles((theme) => ({
   root: {
     boxShadow: "0 12px 18px 0 rgba(0,0,0,0.08)",
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "stretch",
   },
-  tabs: {
-    maxWidth: 400,
-    margin: theme.spacing(3),
-  },
-  data: {},
 }));
-
-/**
- * Tabs enum for ideomatic access
- */
-const Tab = {
-  info: "info",
-  objects: "objects",
-  exif: "exif",
-};
-
-/**
- * Select data-presentation panel
- */
-function dataComponent(tab) {
-  switch (tab) {
-    case Tab.info:
-      return FileInfoPanel;
-    case Tab.objects:
-      return ObjectsPanel;
-    case Tab.exif:
-      return ExifPanel;
-    default:
-      console.error(`Unknown tab: ${tab}`);
-      return "div";
-  }
-}
 
 function useMessages() {
   const intl = useIntl();
   return {
-    info: intl.formatMessage({ id: "file.tabInfo" }),
-    objects: intl.formatMessage({ id: "file.tabObjects" }),
-    exif: intl.formatMessage({ id: "file.tabExif" }),
     ariaLabel: intl.formatMessage({ id: "aria.label.fileAttributesRegion" }),
   };
 }
@@ -67,9 +31,6 @@ function VideoInformationPane(props) {
   const { file, onJump, className } = props;
   const classes = useStyles();
   const messages = useMessages();
-  const [tab, setTab] = useState(Tab.info);
-
-  const DataPanel = dataComponent(tab);
 
   return (
     <Paper
@@ -77,12 +38,7 @@ function VideoInformationPane(props) {
       role="region"
       aria-label={messages.ariaLabel}
     >
-      <SelectableTabs value={tab} onChange={setTab} className={classes.tabs}>
-        <SelectableTab label={messages.info} value={Tab.info} />
-        <SelectableTab label={messages.objects} value={Tab.objects} />
-        <SelectableTab label={messages.exif} value={Tab.exif} />
-      </SelectableTabs>
-      <DataPanel file={file} className={classes.data} onJump={onJump} />
+      <VideoInformation file={file} onJump={onJump} />
     </Paper>
   );
 }
